issue/870: reoder imports
