+++ src/power_grid_model/_core/data_handling.py
@@ -8,14 +8,14 @@ Data handling
 
 import numpy as np
 
+from power_grid_model._core.data_types import Dataset, SingleDataset
 from power_grid_model._core.dataset_definitions import ComponentType, DatasetType
+from power_grid_model._core.enum import CalculationType, ComponentAttributeFilterOptions
+from power_grid_model._core.errors import PowerGridUnreachableHitError
 from power_grid_model._core.power_grid_dataset import CConstDataset, CMutableDataset
 from power_grid_model._core.power_grid_meta import initialize_array, power_grid_meta_data
-from power_grid_model._utils import process_data_filter
-from power_grid_model.data_types import Dataset, SingleDataset
-from power_grid_model.enum import CalculationType, ComponentAttributeFilterOptions
-from power_grid_model.errors import PowerGridUnreachableHitError
-from power_grid_model.typing import ComponentAttributeMapping
+from power_grid_model._core.typing import ComponentAttributeMapping
+from power_grid_model._core.utils import process_data_filter
 
 
 def get_output_type(*, calculation_type: CalculationType, symmetric: bool) -> DatasetType:
