+++ src/power_grid_model/_core/power_grid_model.py
@@ -18,25 +18,25 @@ from power_grid_model._core.data_handling import (
     prepare_output_view,
     prepare_update_view,
 )
+from power_grid_model._core.data_types import Dataset, SingleDataset
 from power_grid_model._core.dataset_definitions import (
     ComponentType,
     ComponentTypeLike,
     _map_to_component_types,
     _str_to_component_type,
 )
-from power_grid_model._core.error_handling import PowerGridBatchError, assert_no_error, handle_errors
-from power_grid_model._core.index_integer import IdNp, IdxNp
-from power_grid_model._core.options import Options
-from power_grid_model._core.power_grid_core import ConstDatasetPtr, IDPtr, IdxPtr, ModelPtr, power_grid_core as pgc
-from power_grid_model.data_types import Dataset, SingleDataset
-from power_grid_model.enum import (
+from power_grid_model._core.enum import (
     CalculationMethod,
     CalculationType,
     ShortCircuitVoltageScaling,
     TapChangingStrategy,
     _ExperimentalFeatures,
 )
-from power_grid_model.typing import ComponentAttributeMapping
+from power_grid_model._core.error_handling import PowerGridBatchError, assert_no_error, handle_errors
+from power_grid_model._core.index_integer import IdNp, IdxNp
+from power_grid_model._core.options import Options
+from power_grid_model._core.power_grid_core import ConstDatasetPtr, IDPtr, IdxPtr, ModelPtr, power_grid_core as pgc
+from power_grid_model._core.typing import ComponentAttributeMapping
 
 
 class PowerGridModel:
