+++ src/power_grid_model/_core/serialization.py
@@ -10,8 +10,10 @@ from abc import ABC, abstractmethod
 from ctypes import byref
 from enum import IntEnum
 
+from power_grid_model._core.data_types import Dataset
 from power_grid_model._core.dataset_definitions import DatasetType, _map_to_component_types, _str_to_datatype
 from power_grid_model._core.error_handling import assert_no_error
+from power_grid_model._core.errors import PowerGridSerializationError
 from power_grid_model._core.index_integer import IdxC
 from power_grid_model._core.power_grid_core import (
     CharPtr,
@@ -21,9 +23,7 @@ from power_grid_model._core.power_grid_core import (
     power_grid_core as pgc,
 )
 from power_grid_model._core.power_grid_dataset import CConstDataset, CWritableDataset
-from power_grid_model.data_types import Dataset
-from power_grid_model.errors import PowerGridSerializationError
-from power_grid_model.typing import ComponentAttributeMapping
+from power_grid_model._core.typing import ComponentAttributeMapping
 
 
 class SerializationType(IntEnum):
