+++ src/power_grid_model/typing.py
@@ -19,30 +19,4 @@ from power_grid_model._core.power_grid_meta import (  # pylint: disable=unused-i
     DatasetMetaData,
     PowerGridMetaData,
 )
-from power_grid_model.enum import ComponentAttributeFilterOptions
-
-_ComponentAttributeMappingDict = dict[ComponentType, set[str] | list[str] | None | ComponentAttributeFilterOptions]
-
-ComponentAttributeMapping = (
-    set[ComponentTypeVar]
-    | list[ComponentTypeVar]
-    | ComponentAttributeFilterOptions
-    | None
-    | _ComponentAttributeMappingDict
-)
-"""
-Type hint for mapping component attributes.
-
-`ComponentAttributeMapping` can be one of the following:
-
-- A set of :class:`ComponentType` or `str`
-
-- A list of :class:`ComponentType` or `str`
-
-- A :class:`ComponentAttributeFilterOptions <power_grid_model.enum.ComponentAttributeFilterOptions>` value
-
-- `None`
-
-- A dictionary mapping :class:`ComponentType` to a set, list, `None` or
-  :class:`ComponentAttributeFilterOptions <power_grid_model.enum.ComponentAttributeFilterOptions>`
-"""
+from power_grid_model._core.typing import ComponentAttributeMapping  # pylint: disable=unused-import
