+++ tests/unit/test_0Z_model_validation.py
@@ -9,7 +9,7 @@ from typing import Callable
 import numpy as np
 import pytest
 
-from power_grid_model._utils import convert_batch_dataset_to_batch_list
+from power_grid_model._core.utils import convert_batch_dataset_to_batch_list
 from power_grid_model.enum import TapChangingStrategy
 
 from .utils import EXPORT_OUTPUT, PowerGridModelWithExt, compare_result, import_case_data, pytest_cases, save_json_data
